backend/foodgram/requirements.txt
Django==3.2.18
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2-binary==2.8.6
PyJWT==2.4.0
python-dotenv==0.21.0
Pillow==9.3.0
webcolors==1.11.1
djoser==2.0.1
django-utils-six==2.0
gunicorn==20.0.4
django-colorfield==0.8.0